Move "migrate()" under "/api/v2/users/<username>/"
