+++ kpi/management/commands/sync_kobocat_xforms.py
@@ -14,10 +14,10 @@ from django.core.exceptions import ImproperlyConfigured
 from django.core.management import call_command
 from django.core.management.base import BaseCommand
 from django.db import transaction
+from formpack.utils.xls_to_ss_structure import xls_to_dicts
 from pyxform import xls2json_backends
 from rest_framework.authtoken.models import Token
 
-from formpack.utils.xls_to_ss_structure import xls_to_dicts
 from kpi.constants import PERM_FROM_KC_ONLY
 from kpi.utils.log import logging
 from kpi.deployment_backends.kc_access.shadow_models import (
